a bit of work has been done...
git-svn-id: svn://projects.prophetsway.net/trunk/ProphetsWay.Utilities@21 e1357c03-9c8d-804b-bc24-f4024500f409

+++ Logger.cs
@@ -95,12 +95,30 @@ namespace ProphetsWay.Utilities
 		public FileDestination(string fileName, LogLevels level = LogLevels.Debug, bool clearFile = true)
 		{
 			_file = new FileInfo(fileName);
-
-			if (clearFile)
-				_file.Delete();
-
 			_logLevel = level;
-			_writer = _file.AppendText();
+
+			try
+			{
+				if (_file.Exists && !clearFile)
+				{
+					_writer = _file.AppendText();
+				}
+				else
+				{
+					if (_file.Exists)
+						_file.Delete();
+
+					if (_file.Directory != null && !_file.Directory.Exists)
+						_file.Directory.Create();
+
+					_writer = _file.CreateText();
+				}
+			}
+			catch (Exception ex)
+			{
+				Console.WriteLine(ex.Message);
+				throw;
+			}
 		}
 
 		private readonly LogLevels _logLevel;
